# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Olimpiadas",
        "description": "Documentación de la API para evaluaciones y evaluadores",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/evaluador/evaluaciones": {
            "get": {
                "tags": [
                    "Evaluador"
                ],
                "summary": "Obtener evaluaciones de un evaluador",
                "operationId": "8f435d68e9ce74ef83514f415694e64c",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "busqueda",
                        "in": "query",
                        "description": "Filtrar evaluaciones por nombre o CI",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evaluaciones obtenidas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_evaluacion": {
                                                        "type": "integer"
                                                    },
                                                    "id_inscrito": {
                                                        "type": "integer"
                                                    },
                                                    "ci": {
                                                        "type": "string"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "area": {
                                                        "type": "string"
                                                    },
                                                    "nivel": {
                                                        "type": "string"
                                                    },
                                                    "nota": {
                                                        "type": "number",
                                                        "nullable": true
                                                    },
                                                    "conducta": {
                                                        "properties": {
                                                            "respeto": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "integridad": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "puntualidad": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron evaluaciones"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/evaluadores/evaluaciones/{id}": {
            "put": {
                "tags": [
                    "Evaluador"
                ],
                "summary": "Actualiza una evaluación existente",
                "description": "Permite actualizar la nota, descripción y los aspectos de conducta de una evaluación. Solo se envían los campos que se desean actualizar.",
                "operationId": "updateEvaluacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la evaluación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nota": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Buen desempeño general"
                                    },
                                    "conducta": {
                                        "properties": {
                                            "respeto": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "integridad": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "puntualidad": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evaluación actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nota": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 85
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Buen desempeño general"
                                                    },
                                                    "conducta": {
                                                        "properties": {
                                                            "respeto": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "integridad": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "puntualidad": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "estado_clasificacion": {
                                                        "type": "string",
                                                        "example": "clasificado"
                                                    },
                                                    "id_inscripcion": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "id_fase": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-18T11:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-18T11:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evaluación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró la evaluación."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ServerError"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}